requirements.txt
requests
beautifulsoup4
pdfplumber
python-telegram-bot